LayoutTests/http/tests/webgpu/webgpu/shader/validation/decl/var_access_mode-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Query `webgpu:shader,validation,decl,var_access_mode:*` does not match any cases

Harness Error (FAIL), message = Unhandled rejection: Query `webgpu:shader,validation,decl,var_access_mode:*` does not match any cases
